refactor: separate the template config file related tests to an own folder
