+++ tests/integration/helpers/project_builder.rs
@@ -1,3 +1,4 @@
+use indoc::indoc;
 use std::fs;
 use std::io::Write;
 use std::str;
@@ -33,15 +34,20 @@ impl ProjectBuilder {
     /// - one file `Cargo.toml` in it
     /// - one placeholder `project-name`
     pub fn init_default_template(self) -> Self {
+        self.default_manifest().init_git()
+    }
+
+    /// creates a `Cargo.toml` manifest with a `project-name` placeholder
+    pub fn default_manifest(self) -> Self {
         self.file(
             "Cargo.toml",
-            r#"[package]
-name = "{{project-name}}"
-description = "A wonderful project"
-version = "0.1.0"
-"#,
+            indoc! {r#"
+                [package]
+                name = "{{project-name}}"
+                description = "A wonderful project"
+                version = "0.1.0"
+            "#},
         )
-        .init_git()
     }
 
     pub fn file(mut self, name: &str, contents: impl AsRef<str>) -> Self {
