+++ tests/integration/template_config_file/mod.rs
@@ -1,0 +1,3 @@
+mod includes_and_excludes;
+mod misc;
+mod placeholders;
